#7 code refactored
